(PC-42398)[API] feat: move jwt utils from core.users.utils to utils.jwt
